requirements.txt
# Web framework and server
fastapi==0.103.1
uvicorn==0.23.2

# Machine learning
tensorflow==2.12.0
numpy>=1.22.0,<1.24.0  

# Image processing
pillow==10.0.1

# File handling and templates
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1